python/requirements.txt
twine
exfunc>=0.5.1
flake8
langchain==0.3.4
langchain-openai==0.2.2
mypy==1.7.0
pydantic==2.9.2
pyright==1.1.350
python-dotenv==1.0.1
ruff==0.4.4
